tgbot/requirements.txt
aiopg
aiohttp
aiogram
python-dotenv
psycopg2-binary